streamline commandbar menu item creation
