+++ javascript/app-framework/src/components/app/ResourcesPane/FileTreeCommands.hooks.ts
@@ -12,18 +12,22 @@ import { useDataPackage } from '../../../hooks/useDataPackage.js'
 import { usePersistenceService } from '../../../hooks/usePersistenceService.js'
 import type { ProfilePlugin } from '../../../types.js'
 import { CommandBarSection } from '../../../types.js'
-import {
-	createCommandBar,
-	openMenuItems,
-	saveMenuItems,
-} from './FileTreeCommands.utils.js'
 import { TABLE_TYPES, ZIP_TYPES } from './ResourcesPane.constants.js'
+import { icons } from './ResourcesPane.styles.js'
 import type { FileDefinition } from './ResourcesPane.types.js'
 
+/**
+ * Gets the file-managament commandbar items
+ *
+ * @param examples - The provided examples
+ * @param expanded - Whether the pane is expended
+ * @param setFile - Set the imported file
+ * @param plugins - The profile plugins
+ * @returns A set of commands for the file management
+ */
 export function useFileManagementCommands(
 	examples: FileDefinition[],
 	expanded: boolean,
-	onOpenFileRequested: (acceptedFiles: string[]) => Promise<BaseFile>,
 	setFile: (file: BaseFile) => void,
 	plugins: Map<string, ProfilePlugin>,
 ): {
@@ -32,12 +36,153 @@ export function useFileManagementCommands(
 	onSaveCommands: IContextualMenuItem[]
 	newCommands: IContextualMenuItem[]
 } {
-	const uploadZip = useUploadZip()
-	const onClickDownloadZip = useDownloadZip()
 	const dataPackage = useDataPackage()
 	const isDataPackageEmpty = useObservableState(
 		dataPackage.isEmpty$,
 		dataPackage.isEmpty,
+	)
+	const newCommands = useNewMenuCommands(plugins)
+	const saveCommands = useSaveMenuCommands(plugins)
+	const openCommands = useOpenMenuItems(examples, plugins, setFile)
+	const theme = useTheme()
+
+	const commands = useMemo<ICommandBarItemProps[]>(
+		() => [
+			{
+				key: 'new',
+				text: expanded ? 'New' : '',
+				iconProps: icons.newFile,
+				subMenuProps: {
+					items: newCommands,
+				},
+				buttonStyles: {
+					root: {
+						background: theme.palette.neutralLighter,
+					},
+				},
+			},
+			{
+				key: 'open',
+				text: expanded ? 'Open' : undefined,
+				iconProps: icons.openFile,
+				iconOnly: !expanded,
+				subMenuProps: {
+					items: openCommands,
+				},
+				buttonStyles: {
+					root: {
+						background: theme.palette.neutralLighter,
+					},
+				},
+			},
+			{
+				key: 'save',
+				text: expanded ? 'Save' : undefined,
+				iconProps: icons.save,
+				iconOnly: !expanded,
+				disabled: isDataPackageEmpty,
+				subMenuProps: {
+					items: saveCommands,
+				},
+				buttonStyles: {
+					root: {
+						background: theme.palette.neutralLighter,
+					},
+				},
+			},
+		],
+		[
+			theme,
+			newCommands,
+			openCommands,
+			saveCommands,
+			expanded,
+			isDataPackageEmpty,
+		],
+	)
+
+	return {
+		commands,
+		newCommands,
+		onOpenCommands: openCommands,
+		onSaveCommands: saveCommands,
+	}
+}
+
+/**
+ * Gets a list of menu items for the 'New' Menu
+ * @param plugins - The profile plugins
+ * @returns
+ */
+function useNewMenuCommands(
+	plugins: Map<string, ProfilePlugin>,
+): IContextualMenuItem[] {
+	return useMemo<IContextualMenuItem[]>(
+		() => getPluginCommands(CommandBarSection.New, plugins),
+		[plugins],
+	)
+}
+
+/**
+ * Gets a list of menu items for the 'Save' Menu
+ * @param plugins - The profile plugins
+ * @returns
+ */
+function useSaveMenuCommands(
+	plugins: Map<string, ProfilePlugin>,
+): IContextualMenuItem[] {
+	const onClickDownloadZip = useDownloadZip()
+	return useMemo<IContextualMenuItem[]>(() => {
+		const result: IContextualMenuItem[] = [
+			{
+				key: 'project',
+				text: 'Project',
+				iconProps: icons.project,
+				onClick: onClickDownloadZip,
+			},
+		]
+		result.push(...getPluginCommands(CommandBarSection.Save, plugins))
+		return result
+	}, [onClickDownloadZip])
+}
+
+function useDownloadZip(): () => void {
+	const persistence = usePersistenceService()
+	return useCallback(
+		() =>
+			void persistence
+				.save()
+				.catch(err => console.error('error loading project', err)),
+		[persistence],
+	)
+}
+
+/**
+ * Gets a list of menu items for the 'Open' Menu
+ * @param plugins - The profile plugins
+ * @returns
+ */
+function useOpenMenuItems(
+	examples: FileDefinition[],
+	plugins: Map<string, ProfilePlugin>,
+	setFile: (file: BaseFile) => void,
+): IContextualMenuItem[] {
+	const dataPackage = useDataPackage()
+	const uploadZip = useUploadZip()
+	const onOpenFileRequested = useOnOpenFileRequested()
+
+	const onClickExample = useCallback(
+		(ex: FileDefinition) => {
+			void fetch(ex.url)
+				.then(res => res.blob())
+				.then(data => {
+					const files = new Map<string, Blob>()
+					files.set('datapackage.json', data)
+					return dataPackage.load(files)
+				})
+				.catch(err => console.error('error loading example file', err))
+		},
+		[dataPackage],
 	)
 
 	const onClickUploadTable = useCallback(
@@ -56,69 +201,37 @@ export function useFileManagementCommands(
 		[uploadZip, onOpenFileRequested],
 	)
 
-	const onClickExample = useCallback(
-		(ex: FileDefinition) => {
-			void fetch(ex.url)
-				.then(res => res.blob())
-				.then(data => {
-					const files = new Map<string, Blob>()
-					files.set('datapackage.json', data)
-					return dataPackage.load(files)
-				})
-				.catch(err => console.error('error loading example file', err))
-		},
-		[dataPackage],
-	)
-
-	const newCommands = useMemo<IContextualMenuItem[]>(
-		() => getPluginCommands(CommandBarSection.New, plugins),
-		[plugins],
-	)
-
-	const openCommands = useMemo(() => {
-		const result = openMenuItems(
-			examples,
-			onClickExample,
-			onClickUploadTable,
-			onClickUploadZip,
-		)
-		result.push(...getPluginCommands(CommandBarSection.Open, plugins))
+	return useMemo<IContextualMenuItem[]>(() => {
+		const result: IContextualMenuItem[] = [
+			{
+				key: 'csv',
+				text: 'CSV File',
+				iconProps: icons.table,
+				onClick: onClickUploadTable,
+			},
+			{
+				key: 'zip',
+				text: 'Project',
+				iconProps: icons.project,
+				onClick: onClickUploadZip,
+			},
+		]
+		if (examples.length > 0) {
+			result.push({
+				key: 'examples',
+				text: 'Example',
+				subMenuProps: {
+					items: examples.map(example => ({
+						key: example.name,
+						text: example.name,
+						onClick: () => onClickExample(example),
+					})),
+				},
+			})
+			result.push(...getPluginCommands(CommandBarSection.Open, plugins))
+		}
 		return result
-	}, [examples, plugins, onClickExample, onClickUploadTable, onClickUploadZip])
-
-	const saveCommands = useMemo(() => {
-		const result = saveMenuItems(onClickDownloadZip)
-		result.push(...getPluginCommands(CommandBarSection.Save, plugins))
-		return result
-	}, [plugins, onClickDownloadZip])
-
-	const theme = useTheme()
-	const commands = useMemo<ICommandBarItemProps[]>(
-		() =>
-			createCommandBar(
-				expanded,
-				!isDataPackageEmpty,
-				newCommands,
-				openCommands,
-				saveCommands,
-				theme,
-			),
-		[
-			theme,
-			isDataPackageEmpty,
-			expanded,
-			openCommands,
-			saveCommands,
-			newCommands,
-		],
-	)
-
-	return {
-		commands,
-		newCommands,
-		onOpenCommands: openCommands,
-		onSaveCommands: saveCommands,
-	}
+	}, [examples, plugins, onClickUploadTable, onClickUploadZip, onClickExample])
 }
 
 function useUploadZip(): (file: BaseFile) => void {
@@ -132,18 +245,7 @@ function useUploadZip(): (file: BaseFile) => void {
 	)
 }
 
-function useDownloadZip(): () => void {
-	const persistence = usePersistenceService()
-	return useCallback(
-		() =>
-			void persistence
-				.save()
-				.catch(err => console.error('error loading project', err)),
-		[persistence],
-	)
-}
-
-export function useOnOpenFileRequested(): (
+function useOnOpenFileRequested(): (
 	acceptedFileTypes: string[],
 ) => Promise<BaseFile> {
 	return useCallback((acceptedFileTypes: string[]) => {
